requirements.txt
alembic==1.13.2
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.8.1
asyncmy==0.2.9
Authlib==1.3.1
bcrypt==4.0.1
certifi==2024.7.4
cffi==1.16.0
click==8.1.7
colorama==0.4.6
cryptography==42.0.8
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.2.0
exceptiongroup==1.2.1
fastapi==0.111.0
fastapi-cli==0.0.4
fastapi-pagination==0.12.26
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
inflect==7.3.1
Jinja2==3.1.4
loguru==0.7.2
Mako==1.3.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
more-itertools==10.3.0
orjson==3.10.6
passlib==1.7.4
pyasn1==0.6.0
pycparser==2.22
pydantic==2.8.2
pydantic-settings==2.2.1
pydantic_core==2.20.1
Pygments==2.18.0
PyMySQL==1.1.1
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
PyYAML==6.0.1
redis==5.0.7
rich==13.7.1
rsa==4.9
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sqlacodegen==2.3.0
SQLAlchemy==2.0.31
sqlmodel==0.0.19
starlette==0.37.2
typeguard==4.3.0
typer==0.12.3
typing_extensions==4.11.0
ujson==5.10.0
uvicorn==0.30.1
watchfiles==0.22.0
websockets==12.0
win32-setctime==1.1.0
